+++ src/Aardvark.Geometry.Quadtree/Builder.fs
@@ -6,12 +6,22 @@ open System.Collections.Generic
 
 module Builder =
 
-    let private debugOutput = false
+    let private debugOutput =
+    #if DEBUG
+        true
+    #else
+        false
+    #endif
 
     let rec private build' (sampleExponent : int) (rootCell : Cell2d) (patches : LayerSet[]) =
 
         if debugOutput then
             printfn "[DEBUG] build' rootCell = %A, %d patches" rootCell patches.Length
+
+        for p in patches do
+            invariantm (p.SampleExponent = sampleExponent) 
+                (fun () -> sprintf "Sample exponent %d does not match patch sample exponent %d." sampleExponent p.SampleExponent)
+                "28d1fdd1-2da6-4329-a065-c134c1351ffc"
 
         match patches.Length with
 
@@ -38,13 +48,35 @@ module Builder =
                 if debugOutput then
                     printfn "[DEBUG] MERGE %d patches; %A" n (bbWindow - rootBounds.Min)
 
-                
                 //if debugOutput then
                 //    for patch in patches do
                 //        printfn "[DEBUG]    %A (exact %A)" (patch.SampleWindow - rootBounds.Min) (bbWindow - rootBounds.Min)
 
+                // adjust root cell for split limit
+                let requiredRootCellSplitLimit = sampleExponent + BuildConfig.Default.SplitLimitPowerOfTwo
+
+                let mutable rootCell = rootCell
+
+                if requiredRootCellSplitLimit <> rootCell.Exponent then
+
+                    invariantm (rootCell.Exponent < requiredRootCellSplitLimit) 
+                        (fun () -> sprintf "Expected root cell exponent %d to be smaller than requiredRootCellSplitLimit %d." rootCell.Exponent requiredRootCellSplitLimit)
+                        "4911adf3-7b87-4234-9bcc-bc3076df846e"
+
+                    if debugOutput then
+                        printfn "[DEBUG] must adjust root cell %A exponent to %d" rootCell requiredRootCellSplitLimit
+
+                    while rootCell.Exponent < requiredRootCellSplitLimit do rootCell <- rootCell.Parent
+
+                    if debugOutput then
+                        printfn "[DEBUG] adjusted root cell is %A" rootCell
+
                 let merged = LayerSet.Merge patches
                 let qnode = QNode(Guid.NewGuid(), ebb, rootCell, BuildConfig.Default.SplitLimitPowerOfTwo, merged)
+                
+                if debugOutput then
+                    printfn "[DEBUG] CREATED QNode with split limit = %d" qnode.SplitLimitExponent
+
                 let xs = qnode |> InMemoryNode |> Query.All Query.Config.Default |> Seq.map (fun x -> x.GetSamples<V4f>(Defs.HeightsBilinear4f)) |> Array.ofSeq |> Array.collect id
                 //for x in xs do printfn "%A" x
                 qnode |> InMemoryNode
@@ -142,4 +174,16 @@ type Builder () =
             ) 
             None // initial state
         
-        
+    member this.Print () : unit =
+
+        printfn "Builder("
+        let ps = patches |> Seq.sortByDescending (fun kv -> kv.Key) |> Seq.map (fun kv -> kv.Value) |> Seq.collect id
+        for p in ps do
+          printfn "  exp = %5d  bb = %A" p.SampleExponent p.BoundingBox
+
+        let bbGlobal = Box2d(ps |> Seq.map (fun p -> p.BoundingBox))
+        printfn "  global bb        = %A" bbGlobal
+        printfn "  global root cell = %A" (Cell2d bbGlobal)
+
+        printfn "  )"
+        ()